Cast u32 eta to f64

+++ main.rs
@@ -30,7 +30,7 @@ fn crossing(a: Vec<f64>, b: Vec<f64>, data: Vec<Vec<f64>>, ndt: Vec<f64>, eta: u
   let b_index_usize = b_index as usize;
   let a_ndt = ndt[a_index_usize];
   let b_ndt = ndt[b_index_usize];
-  let eta_range = 0_f64..(eta - 1_f64);
+  let eta_range = 0_f64..((eta as f64) - 1_f64);
 }
 
 fn alongLine(a: Vec<f64>, b: Vec<f64>, how_far: f64): Vec<f64> {
